feat(agentic-ai): support Google Vertex AI models
