+++ connectors/agentic-ai/pom.xml
@@ -80,6 +80,10 @@
       <groupId>dev.langchain4j</groupId>
       <artifactId>langchain4j-open-ai</artifactId>
     </dependency>
+    <dependency>
+      <groupId>dev.langchain4j</groupId>
+      <artifactId>langchain4j-vertex-ai-gemini</artifactId>
+    </dependency>
 
     <dependency>
       <groupId>com.azure</groupId>
